Don't wrap updates in counter_culture_active
This is an artifact of updates running after commit by default, which is
no longer the default, or in fact even tested. It causes to problems as
demonstrated by the new (formerly failing) test case.

This also deprecates execute_after_commit, as it is currently untested
and there is no real reason to use it anymore.