Use a MAX() aggregation when fetching the current count during reconciliation
On PostgreSQL, if the table where the counts are stored doesn't have a primary_key specified (in our case, because it's actually a view), reconciliation fails with an error:

`must appear in the GROUP BY clause or be used in an aggregate function`

This is because, on Postgres:

> When GROUP BY is present, or any aggregate functions are present, it is not valid for the SELECT list expressions to refer to ungrouped columns except within aggregate functions or when the ungrouped column is functionally dependent on the grouped columns, since there would otherwise be more than one possible value to return for an ungrouped column. A functional dependency exists if the grouped columns (or a subset thereof) are the primary key of the table containing the ungrouped column.

When there's no primary key index present, even though Rails knows what the primary key is, no functional dependency gets detected at the DB level. This MAX agg should effectively be a no-op in every case, but it prevents PG from blowing up. An alternate fix to this same problem would be adding "#{relation_class_table_name}.#{column_name}" to the GROUP BY fields.

Note: the specs added in this commit will pass on SQLite even without these fixes — the fail and are fixed as expected when running the specs on a PostgreSQL DB